@@ -4,37 +4,44 @@
   "description": "webpack demo",
   "main": "app.js",
   "scripts": {
-    "start": "webpack-dev-server --inline -d & open http://localhost:8080/webpack-dev-server/index.html"
+    "_start_comment1": "webpack-dev-server -d & open http://localhost:8080/webpack-dev-server/index.html",
+    "_start_comment": "webpack-dev-server --inline -d & open http://localhost:8080/webpack-dev-server/index.html",
+    "start": "NODE_ENV=development node dev-server.js & open http://localhost:8080/webpack-dev-server/index.html",
+    "build": "rimraf build && NODE_ENV=production webpack"
   },
   "author": "roshan",
   "license": "ISC",
   "devDependencies": {
     "autoprefixer-loader": "^3.2.0",
-    "babel-core": "^6.8.0",
-    "babel-loader": "^6.2.4",
-    "babel-preset-es2015": "^6.6.0",
-    "babel-preset-react": "^6.5.0",
+    "babel-core": "^6.22.1",
+    "babel-loader": "^7.0.0",
+    "babel-preset-es2015": "^6.22.0",
+    "babel-preset-react": "^6.22.0",
+    "babel-preset-stage-0": "^6.24.1",
     "compass-mixins": "^0.12.10",
-    "css-loader": "^0.23.1",
-    "eslint": "^2.9.0",
+    "copy-webpack-plugin": "^4.0.1",
+    "css-loader": "^0.28.2",
+    "eslint": "^3.19.0",
     "eslint-loader": "^1.3.0",
-    "eslint-plugin-react": "^5.0.1",
-    "file-loader": "^0.8.5",
+    "eslint-plugin-react": "^7.0.1",
+    "extract-text-plus-webpack-plugin": "^1.0.0",
+    "file-loader": "^0.11.1",
+    "flux": "^3.1.2",
     "jshint": "^2.9.2",
     "jshint-loader": "^0.8.3",
-    "flux": "^2.1.1",
-    "node-sass": "^3.7.0",
-    "react-dom": "^15.1.0",
-    "react-router": "^2.5.2",
-    "sass-loader": "^3.2.0",
+    "node-sass": "^4.5.3",
+    "react-dom": "^15.4.2",
+    "react-router": "^4.1.1",
+    "rimraf": "^2.6.1",
+    "sass-loader": "^6.0.5",
     "strip-loader": "^0.1.2",
-    "style-loader": "^0.13.1",
+    "style-loader": "^0.18.1",
     "url-loader": "^0.5.7",
-    "webpack": "^1.13.0",
-    "webpack-dev-server": "^1.14.1"
+    "webpack": "^2.6.0",
+    "webpack-dev-server": "^2.4.5"
   },
   "dependencies": {
-    "lodash": "^4.13.1",
-    "react": "^15.1.0"
+    "lodash": "^4.17.4",
+    "react": "^15.4.2"
   }
 }
