@@ -1,2 +1,8 @@
 
-console.log('logging from the utils.js file');
+// console.log('logging from the utils.js file');
+
+
+
+if(module.hot){
+	module.hot.accept();
+}
